Add ugly hack for clearing media and view cache in thumbnail component
Problem is that Kwc_Basic_ImageParent_Events allows to freely choose the image component but
the events suppose that the creating class is the image component. Here the _getImageComponent()
method is overwritten and causing problems when deleting the cache.
Solution: Kwc_Basic_ImageParent_Component only show parent image (changed in 4.2)